app/woodAnomaly/requirements.txt
pandas
numpy
tensorflow
opencv
streamlit
pillow